requirements.txt
mysql-connector-python==8.4.0
prometheus_client==0.20.0
PyYAML==6.0.1
